+++ src/sources/ossindex.rs
@@ -130,7 +130,7 @@ impl OssIndexSource {
             .connect_timeout(CONNECT_TIMEOUT)
             .build()
             .unwrap_or_default();
-        
+
         let retry_policy = ExponentialBackoff::builder().build_with_max_retries(3);
         let client = ClientBuilder::new(raw_client)
             .with(RetryTransientMiddleware::new_with_policy(retry_policy))